requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.4.1
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==42.0.2
cymem==2.0.8
decorator==5.1.1
distro==1.9.0
Django==3.2.23
djangorestframework==3.14.0
dnspython==2.6.1
docxcompose==1.4.0
docxtpl==0.16.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
geocoder==1.38.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.6
intel-openmp==2021.4.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.63
language_data==1.2.0
loguru==0.7.2
lxml==5.1.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
openai==1.30.3
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pika==1.3.2
preshed==3.0.9
probableparsing==0.0.1
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
PyMySQL==1.1.0
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
ratelim==0.1.6
regex==2024.5.10
requests==2.31.0
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
thinc==8.2.3
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.2
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.0
usaddress==0.5.10
usaddress-scourgify==0.6.0
wasabi==1.1.2
weasel==0.3.4
win32-setctime==1.1.0
yaml-config==0.1.5
yarl==1.9.4